feat(steering): add model linking for SAEs and disable cache consistently
Major Changes:
- Add model_id field to SAE download to link SAEs with local models for steering
- Disable KV cache for both steered and unsteered generation for consistency
- Fix SAE download path to use actual nested path (e.g., layer_20/width_16k/average_l0_71)
- Improve SAE file sorting: layer → dimension → L0 value (natural order)

SAE Management:
- Add model selector dropdown in DownloadFromHF component (required field)
- Store model_id in external_saes table for steering lookups
- Fix local_path storage to point to actual SAE files, not root directory

Steering Service:
- Add disable_cache=True to unsteered generation in generate_comparison
- Add disable_cache=True to both generations in generate_strength_sweep
- Ensures consistent behavior between steered/unsteered outputs

Extraction Enhancements:
- Add external_sae_id support for extraction jobs
- Create StartExtractionModal component for external SAE extraction
- Refactor extraction service with improved error handling
- Add new API endpoints for SAE-based feature extraction

+++ backend/src/ml/sparse_autoencoder.py
@@ -568,7 +568,8 @@ class JumpReLUFunction(Function):
             Activated features with threshold gating
         """
         # Heaviside step function: H(z - θ) = 1 if z > θ, else 0
-        gate = (z > threshold).float()
+        # Use z.dtype to preserve half precision (float16) when model is in half mode
+        gate = (z > threshold).to(z.dtype)
 
         # JumpReLU output: z ⊙ H(z - θ)
         output = z * gate
